pprl-core/pprl-core-encoder/src/test/java/de/unileipzig/dbs/pprl/core/encoder/hardening/XorFoldingTest.java
/*
 * Copyright © 2018 - 2021 Leipzig University (Database Research Group)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package de.unileipzig.dbs.pprl.core.encoder.hardening;

import de.unileipzig.dbs.pprl.core.common.BitSetUtils;
import de.unileipzig.dbs.pprl.core.common.model.impl.BitSetVector;
import de.unileipzig.dbs.pprl.core.common.model.api.BitVector;
import org.junit.jupiter.api.Test;

import java.util.BitSet;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;

class XorFoldingTest {

  @Test
  void harden() {

    BitSet bs = BitSetUtils.generateRandomBitSet(32, 16);
    BitVector bv = new BitSetVector(32, bs);

    BitVectorHardener bvHardener = new XorFolding();
    BitVector bvH = bvHardener.harden(bv);

    assertEquals(bv.getLength() / 2, bvH.getLength());
    assertTrue(bv.getCardinality() >= bvH.getCardinality());
  }
}